+++ modules/net.plantabyte.drptrace/src/net/plantabyte/drptrace/Tracer.java
@@ -1,6 +1,7 @@
 package net.plantabyte.drptrace;
 
 import net.plantabyte.drptrace.geometry.*;
+import net.plantabyte.drptrace.trace.*;
 
 import java.util.*;
 
@@ -126,231 +127,6 @@ public class Tracer {
 		throw new UnsupportedOperationException("WIP");
 	}
 	
-	private static enum Dir{
-		UP(0), LEFT(1), DOWN(2), RIGHT(3), NONE(4);
-		private byte index;
-		Dir(int index){
-			this.index = (byte)index;
-		}
-		private static Dir[] counterClockwiseArray = {UP,LEFT,DOWN,RIGHT};
-		
-		public Dir rotateCounterClockwise(){
-			return counterClockwiseArray[(this.index + 1) % 4];
-		}
-		public Dir rotateClockwise(){
-			return counterClockwiseArray[(this.index + 3) % 4];
-		}
-	}
-	private static class Corner{
-		public final Vec2i topLeft;
-		public final Vec2i topRight;
-		public final Vec2i bottomLeft;
-		public final Vec2i bottomRight;
-		public final Vec2 center;
-		public Corner(Vec2i topLeft){
-			this(topLeft, topLeft.right(), topLeft.down(), topLeft.down().right());
-		}
-		private Corner(Vec2i topLeft, Vec2i topRight, Vec2i bottomLeft, Vec2i bottomRight){
-			this.topLeft = topLeft;
-			this.topRight = topRight;
-			this.bottomLeft = bottomLeft;
-			this.bottomRight = bottomRight;
-			this.center = midPoint(topLeft, bottomRight);
-		}
-		public Corner left(){
-			return new Corner(this.topLeft.left(), this.topLeft, this.bottomLeft.left(), this.bottomLeft);
-		}
-		
-		public Corner right(){
-			return new Corner(this.topRight, this.topRight.right(), this.bottomRight, this.bottomRight.right());
-		}
-		
-		public Corner up(){
-			return new Corner(this.topLeft.up(), this.topRight.up(), this.topLeft, this.topRight);
-		}
-		
-		public Corner down(){
-			return new Corner(this.bottomLeft, this.bottomRight, this.bottomLeft.down(), this.bottomRight.down());
-		}
-		
-		public Corner move(Dir dir){
-			switch(dir){
-				case UP: return up();
-				case DOWN: return down();
-				case LEFT: return left();
-				case RIGHT: return right();
-				default: return this;
-			}
-		}
-		
-		public Dir dirFrom(Corner origin){
-			if(this.topLeft.y > origin.topLeft.y){
-				return Dir.UP;
-			} else if(this.topLeft.y < origin.topLeft.y){
-				return Dir.DOWN;
-			} else if(this.topLeft.x > origin.topLeft.x){
-				return Dir.RIGHT;
-			} else if(this.topLeft.x < origin.topLeft.x){
-				return Dir.LEFT;
-			} else {
-				return Dir.NONE;
-			}
-		}
-		
-		public Vec2 midpoint(Corner other) {
-			return this.center.add(other.center).mul(0.5);
-		}
-		
-		@Override
-		public boolean equals(final Object o) {
-			if(this == o) {
-				return true;
-			}
-			if(o == null || getClass() != o.getClass()) {
-				return false;
-			}
-			final Corner corner = (Corner) o;
-			return topLeft.equals(corner.topLeft) && topRight.equals(
-					corner.topRight) && bottomLeft.equals(corner.bottomLeft)
-					&& bottomRight.equals(corner.bottomRight);
-		}
-		
-		@Override
-		public int hashCode() {
-			return Objects.hash(topLeft, topRight, bottomLeft, bottomRight);
-		}
-	}
-	private static class TraceMachine{
-		private final IntMap src;
-		private Corner pos;
-		private final Corner initialPos;
-		private Corner oldPos;
-		private final int color;
-		private final List<Vec2> midpoints;
-		public TraceMachine(IntMap source, Vec2i topLeft, int color, List<Vec2> pointTacker){
-			this.src = source;
-			this.pos = new Corner(topLeft);
-			this.initialPos = pos;
-			this.oldPos = pos.down();
-			this.color = color;
-			this.midpoints = pointTacker;
-		}
-		private boolean isColor(Vec2i p){
-			return src.isInRange(p.x, p.y) && src.get(p.x, p.y) == color;
-		}
-		public boolean done(){
-			return this.pos.equals(initialPos);
-		}
-		public void step(){
-			// first, get direction and rotate color checking to match perspective of machine motion
-			var dir = pos.dirFrom(oldPos);
-			var turn = Dir.NONE;
-			int perspective_nl_nr_fl_fr = 0b0000;
-			boolean nl = false, nr = false, fl = false, fr = false;
-			switch(dir){
-				case UP:{
-					nl = isColor(pos.bottomLeft);
-					nr = isColor(pos.bottomRight);
-					fl = isColor(pos.topLeft);
-					fr = isColor(pos.topRight);
-					break;
-				}
-				case DOWN:{
-					nl = isColor(pos.topRight);
-					nr = isColor(pos.topLeft);
-					fl = isColor(pos.bottomRight);
-					fr = isColor(pos.bottomLeft);
-					break;
-				}
-				case LEFT:{
-					nl = isColor(pos.bottomRight);
-					nr = isColor(pos.topRight);
-					fl = isColor(pos.bottomLeft);
-					fr = isColor(pos.topLeft);
-					break;
-				}
-				case RIGHT:{
-					nl = isColor(pos.topLeft);
-					nr = isColor(pos.bottomLeft);
-					fl = isColor(pos.topRight);
-					fr = isColor(pos.bottomRight);
-					break;
-				}
-				default:
-					throw new IllegalStateException("Machine old and current position equal, cannot move");
-			}
-			if(nl) perspective_nl_nr_fl_fr |= 0b1000;
-			if(nr) perspective_nl_nr_fl_fr |= 0b0100;
-			if(fl) perspective_nl_nr_fl_fr |= 0b0010;
-			if(fr) perspective_nl_nr_fl_fr |= 0b0001;
-			// then decide whether to turn or keep going in same direction
-			switch(perspective_nl_nr_fl_fr){
-				case 0b1000:
-					// ..
-					// #.
-					// ^^
-				case 0b0111:
-					// ##
-					// .#
-					// ^^
-				case 0b1001:
-					// .#
-					// #.
-					// ^^
-				case 0b1100:
-					// ..
-					// ##
-					// ^^
-					turn = Dir.LEFT;
-					break;
-				case 0b0100:
-					// ..
-					// .#
-					// ^^
-				case 0b1011:
-					// ##
-					// #.
-					// ^^
-				case 0b0110:
-					// #.
-					// .#
-					// ^^
-				case 0b0011:
-					// ##
-					// ..
-					// ^^
-					turn = Dir.RIGHT;
-					break;
-				case 0b1111:
-					// ##
-					// ##
-					// ^^
-					throw new IllegalStateException("Not on corner, all four are color");
-				case 0b0000:
-					// ..
-					// ..
-					// ^^
-					throw new IllegalStateException("Not on corner, all four are off-color");
-			}
-			Corner newPos;
-			if(turn == Dir.NONE){
-				// keep going straight
-				newPos = pos.move(dir);
-			} else if( turn == Dir.LEFT){
-				// turn left, aka counter-clockwise
-				newPos = pos.move(dir.rotateCounterClockwise());
-			} else {
-				// turn right, aka clockwise
-				newPos = pos.move(dir.rotateClockwise());
-			}
-			// now add edge point from step
-			midpoints.add(pos.midpoint(newPos));
-			// finally, update the position
-			System.out.println(oldPos.topLeft + " -> " + pos.topLeft + " ("+ dir +")" + "\t0b"+Integer.toBinaryString(perspective_nl_nr_fl_fr)); // TODO: remove
-			oldPos = pos;
-			pos = newPos;
-		}
-	}
 	public List<Vec2> followEdge(final IntMap source, final int x, final int y){ // TODO: make private
 		// trace counter-clockwise around the edge
 		final int color = source.get(x,y);
@@ -373,58 +149,4 @@ public class Tracer {
 		throw new UnsupportedOperationException("WIP");
 	}
 	
-	/**
-	 * Gets up, left, down, right neighbors in counter-clockwise order, ending at
-	 * <code>end</code>
-	 * @param center The point to rotate around
-	 * @param end Where the counter-clockwise sequence should end
-	 * @return An array of up, left, down, right neighbors, where the last index
-	 * (3) is equal to end
-	 */
-	private static Vec2i[] neighborsCounterClockwise(Vec2i center, Vec2i end){
-		Vec2i[] cc = new Vec2i[4];
-		Vec2i[] out = new Vec2i[4];
-		cc[0] = center.up();
-		cc[1] = center.left();
-		cc[2] = center.down();
-		cc[3] = center.right();
-		int i = 0;
-		for(i = 0; i < 3; i++){
-			if(cc[i].equals(end)) break;
-		}
-		for(int n = 0; n < 4; n++){
-			out[n] = cc[(n+i+1)%4];
-		}
-		return out;
-	}
-	
-	private static Vec2i nextNeighborCounterClockwise(Vec2i center, Vec2i from){
-		Vec2i[] cc = new Vec2i[4];
-		cc[0] = center.up();
-		cc[1] = center.left();
-		cc[2] = center.down();
-		cc[3] = center.right();
-		int i = 0;
-		for(i = 0; i < 3; i++){
-			if(cc[i].equals(from)) break;
-		}
-		return cc[(i+3)%4];
-	}
-	
-	private static Vec2i nextNeighborClockwise(Vec2i center, Vec2i from){
-		Vec2i[] cc = new Vec2i[4];
-		cc[0] = center.up();
-		cc[1] = center.right();
-		cc[2] = center.down();
-		cc[3] = center.left();
-		int i = 0;
-		for(i = 0; i < 3; i++){
-			if(cc[i].equals(from)) break;
-		}
-		return cc[(i+3)%4];
-	}
-	
-	private static Vec2 midPoint(Vec2i a, Vec2i b){
-		return new Vec2(0.5*(a.x+b.x), 0.5*(a.y+b.y));
-	}
 }
